refactor: PlaygroundService와 ProjectService 인터페이스 분리
